Loïc_FOTSO_Léo_Bleibel/requirements.txt
python-dotenv==1.0.1
openai==1.61.0
duckduckgo_search==7.3.0
streamlit==1.42.0
requests==2.32.3
beautifulsoup4==4.12.3